fix: conflict resolution 2

+++ app/src/main/java/com/mishaismenska/hackatonrsschoolapp/data/AppDataRepositoryImpl.kt
@@ -2,11 +2,12 @@ package com.mishaismenska.hackatonrsschoolapp.data
 
 import android.content.Context
 import android.icu.util.Measure
+import androidx.preference.PreferenceManager
 import com.mishaismenska.hackatonrsschoolapp.R
+import com.mishaismenska.hackatonrsschoolapp.data.interfaces.AppDataRepository
 import com.mishaismenska.hackatonrsschoolapp.data.models.DrinkDataModel
 import com.mishaismenska.hackatonrsschoolapp.data.models.UserDataModel
 import com.mishaismenska.hackatonrsschoolapp.data.models.UserWithDrinksDataModel
-import com.mishaismenska.hackatonrsschoolapp.domain.interfaces.AppDataRepository
 import com.mishaismenska.hackatonrsschoolapp.domain.models.DrinkDomainModel
 import com.mishaismenska.hackatonrsschoolapp.staticPresets.Gender
 import java.time.LocalDate
@@ -62,7 +63,7 @@ class AppDataRepositoryImpl @Inject constructor(private val context: Context) :
                 LocalDate.now(),
                 age,
                 gender.ordinal,
-                weight.number as Int,
+                weight.number.toDouble(),
                 weight.unit,
                 context.getString(R.string.default_name)
             )
@@ -78,7 +79,7 @@ class AppDataRepositoryImpl @Inject constructor(private val context: Context) :
         dao.resetDrinks()
     }
 
-    override suspend fun setWeight(newValue: Int) {
+    override suspend fun setWeight(newValue: Double) {
         dao.setWeight(newValue)
     }
